Create AbstractBaseIntegrationTest
