+++ tests/Integration/EndToEnd/CreateAddSourcesCompileExecuteTest.php
@@ -5,24 +5,19 @@ declare(strict_types=1);
 namespace App\Tests\Integration\EndToEnd;
 
 use App\Entity\Job;
-use App\Entity\Test;
-use App\Entity\TestConfiguration;
 use App\Services\JobStore;
-use App\Tests\AbstractBaseFunctionalTest;
+use App\Tests\Integration\AbstractBaseIntegrationTest;
 use App\Tests\Services\BasilFixtureHandler;
 use App\Tests\Services\ClientRequestSender;
 use App\Tests\Services\SourceStoreInitializer;
 use App\Tests\Services\UploadedFileFactory;
-use Doctrine\ORM\EntityManagerInterface;
-use Doctrine\Persistence\ObjectRepository;
 
-class CreateAddSourcesCompileExecuteTest extends AbstractBaseFunctionalTest
+class CreateAddSourcesCompileExecuteTest extends AbstractBaseIntegrationTest
 {
     private ClientRequestSender $clientRequestSender;
     private JobStore $jobStore;
     private UploadedFileFactory $uploadedFileFactory;
     private BasilFixtureHandler $basilFixtureHandler;
-    private EntityManagerInterface $entityManager;
 
     protected function setUp(): void
     {
@@ -51,16 +46,6 @@ class CreateAddSourcesCompileExecuteTest extends AbstractBaseFunctionalTest
         if ($basilFixtureHandler instanceof BasilFixtureHandler) {
             $this->basilFixtureHandler = $basilFixtureHandler;
         }
-
-        $entityManager = self::$container->get(EntityManagerInterface::class);
-        self::assertInstanceOf(EntityManagerInterface::class, $entityManager);
-        if ($entityManager instanceof EntityManagerInterface) {
-            $this->entityManager = $entityManager;
-        }
-
-        $this->removeAllEntities(Job::class);
-        $this->removeAllEntities(Test::class);
-        $this->removeAllEntities(TestConfiguration::class);
 
         $this->initializeSourceStore();
     }
@@ -104,30 +89,5 @@ class CreateAddSourcesCompileExecuteTest extends AbstractBaseFunctionalTest
         if ($sourceStoreInitializer instanceof SourceStoreInitializer) {
             $sourceStoreInitializer->initialize();
         }
-    }
-
-    /**
-     * @param class-string $entityClassName
-     */
-    private function removeAllEntities(string $entityClassName): void
-    {
-        $repository = $this->entityManager->getRepository($entityClassName);
-        if ($repository instanceof ObjectRepository) {
-            $entities = $repository->findAll();
-
-            foreach ($entities as $entity) {
-                $this->entityManager->remove($entity);
-                $this->entityManager->flush();
-            }
-        }
-    }
-
-    protected function tearDown(): void
-    {
-        parent::tearDown();
-
-        $this->removeAllEntities(Job::class);
-        $this->removeAllEntities(Test::class);
-        $this->removeAllEntities(TestConfiguration::class);
     }
 }
